+++ src/main/resources/map/css/styles.css
@@ -19,9 +19,83 @@ main {
 
 #map {
     width: 75%;
-    height: 500px;
+    height: calc(100vh - 80px);
     position: relative;
-    float: right;
+}
+
+#left-panel {
+    width: 25%;
+    height: calc(100vh - 80px);
+    overflow-y: auto;
+    background: white;
+    padding: 10px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.panel-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+#place-list {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+#place-list li {
+    margin-bottom: 10px;
+    padding: 10px;
+    background: #f9f9f9;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+}
+
+#pagination {
+    margin: 10px auto;
+    text-align: center;
+}
+
+#pagination a {
+    display: inline-block;
+    margin-right: 10px;
+}
+
+#pagination .on {
+    font-weight: bold;
+    cursor: default;
+    color: #777;
+}
+
+.place-info {
+    margin-top: 10px;
+    padding: 10px;
+    background: #f0f0f0;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+}
+
+#top-left-buttons {
+    position: absolute;
+    top: 10px;
+    left: 10px;
+    z-index: 10;
+    display: flex;
+    gap: 10px;
+}
+
+#top-left-buttons button {
+    padding: 10px 20px;
+    background-color: #007BFF;
+    color: white;
+    border: none;
+    cursor: pointer;
+    border-radius: 5px;
+}
+
+#top-left-buttons button:hover {
+    background-color: #0056b3;
 }
 
 .top-bar {
@@ -102,24 +176,6 @@ main {
     background-color: #cc0000;
 }
 
-#top-left-buttons {
-    display: flex;
-    gap: 10px;
-}
-
-#top-left-buttons button {
-    padding: 10px 20px;
-    background-color: #007BFF;
-    color: white;
-    border: none;
-    cursor: pointer;
-    border-radius: 5px;
-}
-
-#top-left-buttons button:hover {
-    background-color: #0056b3;
-}
-
 .keyword-list {
     position: absolute;
     background: white;
@@ -187,50 +243,4 @@ main {
 #address-search-results {
     max-height: 200px;
     overflow-y: auto;
-}
-
-#left-panel {
-    width: 25%;
-    height: 500px;
-    overflow-y: auto;
-    background: white;
-    padding: 10px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    float: left;
-}
-
-.panel-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-#place-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-#place-list li {
-    margin-bottom: 10px;
-    padding: 10px;
-    background: #f9f9f9;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-}
-
-#pagination {
-    margin: 10px auto;
-    text-align: center;
-}
-
-#pagination a {
-    display: inline-block;
-    margin-right: 10px;
-}
-
-#pagination .on {
-    font-weight: bold;
-    cursor: default;
-    color: #777;
 }
