+++ src/main/resources/map/js/utils.js
@@ -1,0 +1,8 @@
+// utils.js
+export function toRadians(degrees) {
+    return degrees * Math.PI / 180;
+}
+
+export function toDegrees(radians) {
+    return radians * 180 / Math.PI;
+}
